These changes should be stable.  However, I have made a few changes to get ready for planned changes to General Settings and possibly to new User Settings.
The BCC button in EmailPreview has been commented out.  I plan to make it
a General and/or User option.  The PHPMailer language is currently set to
en (English) in EmailSend.php.  Language should be an option in General Settings.

Hopefully I have improved error messages that will be helpful to someone installing
PHPMailer for the first time.
 Modified Files:
 	churchinfo/EmailPreview.php churchinfo/EmailSend.php
 	churchinfo/Include/ReportFunctions.php
-Ed Davis