Update Person.php
I've found a bug in the global code (not in mine).
When a person don't have a family the map the group will crash.
An administrator of our church assign a member to a group without any address (i.e. : without any family).
The map my group will crash.
I corrected the code to get it work.