Added conditional buttons to PersonView.php
Depending on whether the logged in user is displaying their own record in PersonView.php, we add the following two buttons:
- Change Settings
- Change Password

Both these features were previously included in the User Profile section of Header.php, which I am proposing to deprecate from that area in favor of further alignment with the AdminLTE UI standard.  See other components of this PR for further details.