Issue 4055 - Add support for MySQL on non-standard ports (#4071)
* Issue 4055

Added support for MySQL running on non-standard ports. This maintains backward compatibility with previous configurations that do NOT have '$dbPort' defined in the Config.php file.

* Additional function calls

Found some more function calls that were not including "dbPort" in the parameter list. Fixed the function calls to include it.

* Remove unnecessary logic

Removed redundant "if" block and instead used PHP's default-style in the function arguments list.
http://php.net/manual/en/functions.arguments.php#functions.arguments.default

* Apply fixes from StyleCI (#4074)