Fix filtering classifications with parenthesis
closes #5116

According to the datatables api docs, search is configured to treat the "needle string" as a regex string; therefore, search values containing parentheses were being interpreted as regex gropus.

Since we're building a regex string intended to match a complete line, we should escape regex characters in the needle string.

link: https://datatables.net/reference/api/search()