Create a SystemCalendar to display events which are not pinned to any calendar.
Modify client-side logic to provie a mechanism by which to edit eligible SystemCalendar events (i.e. the new Unpinned Events)

This is a mitigation for #4830 in which events created by the legacy events creation form are not pinned to the calendar, and are therefor invisible.

The old editor is deplrecated, and will be removed in a future release, so I'm not adding the "pin" capability to the old editor.

Closes #4830