builds should use composer.lock to ensure dependencies the same as when developed
From https://getcomposer.org/doc/01-basic-usage.md#:~:text=As%20mentioned%20above%2C%20the%20composer,file%20with%20the%20new%20versions,
> As mentioned above, the composer.lock file prevents you from
> automatically getting the latest versions of your dependencies.

This is a good thing, as it ensures that the packaged version is
stable and matches what was being developed.
If someone wants to update dependencies, it should be tracked as
a commit in the repository.