fix: simplify logic to validate redirect
Looking the RFC 7230 (- https://httpwg.org/specs/rfc7230.html#rfc.section.3.2)
I identified that the header field is defined as case-insensitive,
the header to identify if is available to use rewrite is the follow
code:

```php
header('CRM: would redirect');
```
At the point that I wrote again, is using lowercase.

To make the code a bit smallest than the previous, more objective and
also able to ignore the case as was defined into RFC 7230, I used the
function preg_match and a regex to match row by row from CURL response
(m) together with a case insensitive instruction  (i) at the regex.

Reason to this change:

I made a setup using Nginx and identified this because the Nginx is
respecting the case, but the code don't respect the case.

I maintained the header with CRM as uppercase because is more readdable
and fixed the code to be insensitive.

Signed-off-by: Vitor Mattos <1079143+vitormattos@users.noreply.github.com>